# Conflicts:
#	Net14Online/Net14Web/Program.cs

+++ Net14Online/Net14Web/Controllers/BondsController.cs
@@ -5,6 +5,7 @@ using Net14Web.DbStuff.Models.Bonds;
 using Net14Web.DbStuff.Repositories;
 using Net14Web.Models.Bonds;
 using Net14Web.Services;
+using Net14Web.Services.BondServices;
 
 namespace Net14Web.Controllers
 {
@@ -14,26 +15,34 @@ namespace Net14Web.Controllers
         private BondsRepository _bondsRepository;
         private CouponsRepository _couponsRepository;
         private AuthService _authService;
+        private BondPermissions _bondPermissions;
+        private CouponPermissions _couponPermissions;
 
         public BondsController(BondsRepository bondsRepository,
             CouponsRepository couponsRepository,
-            AuthService authService)
+            AuthService authService,
+            BondPermissions bondPermissions,
+            CouponPermissions couponPermissions
+            )
         {
             _bondsRepository = bondsRepository;
             _couponsRepository = couponsRepository;
             _authService = authService;
+            _bondPermissions = bondPermissions;
+            _couponPermissions = couponPermissions;
         }
         public IActionResult Index()
         {
             var userName = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "name")?.Value ?? "Гость";
             var bdBonds = _bondsRepository.GetBonds(10);
             var bondsViewModel = bdBonds
-                .Select(x => new BondsViewModel
+                .Select(dbBond => new BondsViewModel
                 {
-                    Id = x.Id,
-                    Name = x.Name,
-                    Price = x.Price,
-                    OwnerName = x.Owner?.Login ?? "Кто-то"
+                    Id = dbBond.Id,
+                    Name = dbBond.Name,
+                    Price = dbBond.Price,
+                    OwnerName = dbBond.Owner?.Login ?? "Кто-то",
+                    CanDelete = _bondPermissions.CanDelete(dbBond)
                 })
                 .ToList();
             var viewModel = new BondsIndexViewMode()
@@ -72,8 +81,14 @@ namespace Net14Web.Controllers
 
         public IActionResult Remove(int id)
         {
-            _bondsRepository.Delete(id);
+            var dbBond = _bondsRepository.GetByIdWithOwner(id);
 
+            if (!_bondPermissions.CanDelete(dbBond))
+            {
+                throw new Exception("Ты кто такой?");
+            }
+
+            _bondsRepository.Delete(id);
             return RedirectToAction("Index");
         }
 
@@ -90,13 +105,14 @@ namespace Net14Web.Controllers
         {
             var bdCoupons = _couponsRepository.GetCoupons(10);
             var viewModel = bdCoupons
-                .Select(x => new CouponsViewModel
+                .Select(bdCoupon => new CouponsViewModel
                 {
-                    Id = x.Id,
-                    Date = x.Date,
-                    CouponSize = x.CouponSize,
-                    Bond = x.Bond.Name,
-                    OwnerName = x.Bond.Owner?.Login ?? "Кто-то"
+                    Id = bdCoupon.Id,
+                    Date = bdCoupon.Date,
+                    CouponSize = bdCoupon.CouponSize,
+                    Bond = bdCoupon.Bond.Name,
+                    OwnerName = bdCoupon.Bond.Owner?.Login ?? "Кто-то",
+                    CanDelete = _couponPermissions.CanDelete(bdCoupon)
                 }).ToList();
             return View(viewModel);
         }
